jiayue-cloud-functions/temp/phl_property_data_pipeline/requirements.txt
functions_framework==3.8.2
protobuf==6.30.2
pyproj==3.6.1
python-dotenv==1.1.0
Requests==2.32.3
Shapely==2.0.7
